requirements.txt
python-telegram-bot
ipdb
